refactor[#195] profile API 정리
